+++ frontend/src/components/Footer.js
@@ -1,5 +1,5 @@
-import React from 'react';
 import PropTypes from 'prop-types';
+import React from 'react';
 
 import {List} from 'semantic-ui-react'
 
@@ -11,8 +11,10 @@ class AppFooter extends React.Component {
                     <List id="footerList" floated="right" horizontal size="small">
                         <List.Item>{this.props.version}</List.Item>
                         {this.props.footerLink}
-                        <List.Item href={"https://github.com/ccoors/afu-group-trainer"} target={"_blank"}
-                                   rel={"noopener noreferrer"}>Source</List.Item>
+                        <List.Item href={'https://agt.dl5lq.de/doc'} target={'_blank'}
+                                   rel={'noopener noreferrer'}>Dokumentation</List.Item>
+                        <List.Item href={'https://github.com/ccoors/afu-group-trainer'} target={'_blank'}
+                                   rel={'noopener noreferrer'}>Source</List.Item>
                     </List>
                 </div>
             </footer>