+++ frontend/src/components/QuestionRenderer/CompactQuestionRenderer.js
@@ -1,12 +1,12 @@
-import React from 'react';
 import PropTypes from 'prop-types';
-import {Button, Segment} from "semantic-ui-react";
-import {stringToJSX} from "../../util/util";
-import OutdatedQuestionLabel from "./OutdatedQuestionLabel";
+import React from 'react';
+import {Button, Segment} from 'semantic-ui-react';
+import {questionTitle, stringToJSX} from '../../util/util';
+import OutdatedQuestionLabel from './OutdatedQuestionLabel';
 
 class CompactQuestionRenderer extends React.Component {
     renderSegment(nr, answer) {
-        let colon = answer.trim() !== "" ? ":" : "";
+        let colon = answer.trim() !== '' ? ':' : '';
         return <div>
             <strong>{nr}{colon}</strong>&nbsp;
             {stringToJSX(answer)}
@@ -14,17 +14,19 @@ class CompactQuestionRenderer extends React.Component {
     }
 
     render() {
+        const {question} = this.props;
+
         return <Segment>
-            {this.props.question.outdated &&
+            {question.outdated &&
             <OutdatedQuestionLabel/>}
             {this.props.onDelete &&
             <Button circular icon='close' color="red" floated="right"
                     onClick={() => this.props.onDelete(this.props.question.uuid)}/>}
-            <strong>{this.props.question.id}:</strong> {stringToJSX(this.props.question.question)}
-            {this.renderSegment("A", this.props.question.answers[0])}
-            {this.renderSegment("B", this.props.question.answers[1])}
-            {this.renderSegment("C", this.props.question.answers[2])}
-            {this.renderSegment("D", this.props.question.answers[3])}
+            <strong>{questionTitle(question)}:</strong> {stringToJSX(question.question)}
+            {this.renderSegment('A', question.answers[0])}
+            {this.renderSegment('B', question.answers[1])}
+            {this.renderSegment('C', question.answers[2])}
+            {this.renderSegment('D', question.answers[3])}
         </Segment>;
     }
 }