+++ frontend/src/components/QuestionRenderer/DemoQuestionRenderer.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import {generateDummyQuestion} from "../../util/util";
-import CompactQuestionRenderer from "./CompactQuestionRenderer";
-import QuestionRenderer from "./QuestionRenderer";
-import QuestionRORenderer from "./QuestionRORenderer";
+import {generateDummyQuestion} from '../../util/util';
+import CompactQuestionRenderer from './CompactQuestionRenderer';
+import QuestionRenderer from './QuestionRenderer';
+import QuestionRORenderer from './QuestionRORenderer';
 
 class DemoQuestionRenderer extends React.Component {
     render() {
@@ -10,9 +10,11 @@ class DemoQuestionRenderer extends React.Component {
 
         return (
             <div>
-                <CompactQuestionRenderer question={question} onDelete={() => {}}/>
-                <QuestionRenderer question={question} selectedAnswer={2} selectAnswer={() => {}}/>
-                <QuestionRORenderer question={question} correctAnswer={2} selectedAnswer={0} />
+                <CompactQuestionRenderer question={question} onDelete={() => {
+                }}/>
+                <QuestionRenderer question={question} selectedAnswer={2} selectAnswer={() => {
+                }}/>
+                <QuestionRORenderer question={question} correctAnswer={2} selectedAnswer={0}/>
             </div>
         );
     }